提交第二版重构代码

+++ FangPage.WMS.Admin/rolemanage.cs
@@ -1,35 +1,34 @@
 ﻿using System;
 using System.Collections.Generic;
+using FangPage.Common;
 using FangPage.Data;
 using FangPage.MVC;
 using FangPage.WMS.Model;
+using FangPage.WMS.Web;
 
 namespace FangPage.WMS.Admin
 {
-	// Token: 0x02000047 RID: 71
+	// Token: 0x02000051 RID: 81
 	public class rolemanage : SuperController
 	{
-		// Token: 0x060000AC RID: 172 RVA: 0x0000DBA8 File Offset: 0x0000BDA8
-		protected override void View()
+		// Token: 0x060000C3 RID: 195 RVA: 0x0000F614 File Offset: 0x0000D814
+		protected override void Controller()
 		{
-			if (this.ispost)
+			if (this.ispost && this.action == "delete")
 			{
-				if (this.action == "delete")
+				foreach (int num in FPArray.SplitInt(FPRequest.GetString("chkdel")))
 				{
-					foreach (int num in FPUtils.SplitInt(FPRequest.GetString("chkdel")))
+					if (num > 5)
 					{
-						if (num > 5)
-						{
-							DbHelper.ExecuteDelete<RoleInfo>(num);
-						}
+						DbHelper.ExecuteDelete<RoleInfo>(num);
 					}
 				}
+				FPCache.Remove("FP_ROLELIST");
 			}
 			this.rolelist = DbHelper.ExecuteList<RoleInfo>(OrderBy.ASC);
-			base.SaveRightURL();
 		}
 
-		// Token: 0x040000AB RID: 171
+		// Token: 0x040000E2 RID: 226
 		protected List<RoleInfo> rolelist = new List<RoleInfo>();
 	}
 }
